Text2SQL_V2/requirements.txt
flask
pandas
matplotlib
langchain
langchain-openai
langchain-google-genai
python-dotenv
